refactor(freeze-updates): extract dispatcher patcher
Final split of the 4-job freeze-updates module. Move patchDispatcher,
installDispatcherPatching, and initializeFreezeSupport into
freeze/dispatcher-patch.ts. This was the largest sub-module (~165 lines)
that wraps React's useState/useReducer/useTransition/useSyncExternalStore
factories on every dispatcher object the renderer hands out.

The remaining freeze-updates.ts (199 lines) is now just the orchestrator:
collectFiberRoots, fiber-traversal pause/resume, scheduleReactUpdate,
invokeCallbacks, and the public freezeUpdates() entry that wires it all
together.

Final freeze/ shape:
  freeze-updates.ts          199  (was 564, -365)
  freeze/dispatcher-patch.ts 173
  freeze/hook-queue.ts        87
  freeze/context-dependency  73
  freeze/pending-update-chain 63
  freeze/state.ts             60
  freeze/types.ts             52
  total                      707  (was 564, +143 of headers/imports)

The +143 is module boundaries: each sub-module declares its dependencies
explicitly. The orchestrator file is now a clear 'fiber walk + replay'
view of the freeze, with the React-internals interception details
isolated behind narrow interfaces.

34 freeze specs green.
